refactor: 유저 데이터 관리 코드에 DIP 적용
- UserRepository를 인터페이스로, InMemoryUserRepository를 구현체로 변경하여 DIP 적용함
